backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
django-cors-headers>=4.4
psycopg2-binary>=2.9
Pillow>=10.0
reportlab>=4.0
